pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "active-perception-r1"
version = "0.1.0"
description = "verl-based active perception GRPO scaffold for visual self-verification research."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "packaging>=24.0",
    "Pillow>=11.0.0",
]

[project.optional-dependencies]
bench = [
    "torch>=2.7.0",
    "transformers>=4.56.0,<5",
    "datasets>=4.8.0",
    "pyarrow>=23.0.0",
]
train = [
    "torch==2.9.0",
    "torchaudio==2.9.0",
    "torchvision==0.24.0",
    "transformers>=4.56.0,<5",
    "datasets>=4.8.0",
    "pyarrow>=23.0.0",
    "verl==0.7.1",
    "vllm==0.12.0",
    "ray[default]>=2.54.0,<3",
    "pybase64>=1.4.3",
    "accelerate>=1.13.0",
    "gguf>=0.17.1",
    "hydra-core>=1.3.2",
    "peft>=0.18.1",
]
test = [
    "pytest>=8.3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
